Refactor sampling and implement span end.
Implements span end and refactors some of the lifetime related things in
buckets and the store. Now finished SpanData is moved into the store
rather than the store maintaining a reference to it, which would have
gotten nasty quickly.

Not very pleased with this implementation but should allow testing to
continue and performance can be re-evaluated later.